Much smarter more advanced info frame control. It is now dynamic and automatically displays at more appropriate times with an ability priority list. 1. Meteors 2. Elementals 3. Traps
What this means, if health frame is up, and elementals spawn for example, the info frame will change from health frame to aggro frame until elementals spawn, once they all die, it will automatically change back to health frame. it uses a set of conditions to determine if a frame should be shown, only time there is no info frame is if traps elementals and meteors are at 0 (or options are turned off). If any of them are >0 then what frame is displayed will be based on this priority list.