Basic spirit kings mod.
Of note:
This mod has no normal mode timers. or any that do start, will probably be wrong. these timers were for 10 man heroic logs only.
When bosses are defeated, timers will NOT cancel, since we didn't manage to defeat anything.

Hopefully, if i can come across a transcriptor log for normal mode, or they test it again, i'll be able to fix the above issues. it HAS to be transcriptor log. reg logs are woefully inadequate on this fight