tag 5.3.0
I screwed up with the previous 5.2.6 tag, sorry. The correct version number should be 5.3.0 since this is the first version of DBM for WoW 5.3.
Also, the 5.2.6 tag was based on a branch with a new event handling feature which is probably still somewhat buggy, I didn't realize that I were on that branch, sorry.