Add infrastructure to simplify spreading prototypes across files
This is basically just a table that makes it easy to spread a class
definition across multiple files. Moving EnrageTimer to a new file as a
proof of concept because it's the simplest prototype/object we have.

This should hopefully make it easier to add "TimerV2" etc in the future.