Core-Icons: prevent possible nil table while sorting
On very specific conditions, the clearIconTable was being scheduled to be cleared after Icon Set, however, whenever there was a new call to the sorting method, if it happened at the same time as the elapsed scheduling timer (1.5s), it would wipe the table while the sorting was taking place.

The fix attempts to correct the issue by unscheduling the table wipe before sorting.